+++ src/eps/tutorials/ex46.c
@@ -1,0 +1,118 @@
+/*
+   - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+   SLEPc - Scalable Library for Eigenvalue Problem Computations
+   Copyright (c) 2002-2020, Universitat Politecnica de Valencia, Spain
+
+   This file is part of SLEPc.
+   SLEPc is distributed under a 2-clause BSD license (see LICENSE).
+   - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+*/
+
+static char help[] = "Illustrates passing a sparser matrix to build the preconditioner.\n\n"
+  "The command line options are:\n"
+  "  -n <n>, where <n> = number of grid subdivisions in x dimension.\n"
+  "  -m <m>, where <m> = number of grid subdivisions in y dimension.\n\n";
+
+#include <slepceps.h>
+
+int main(int argc,char **argv)
+{
+  Mat            A,A0;            /* operator matrix */
+  EPS            eps;             /* eigenproblem solver context */
+  ST             st;
+  PetscInt       N,n=24,m,Istart,Iend,II,i,j;
+  PetscBool      flag,terse;
+  PetscErrorCode ierr;
+
+  ierr = SlepcInitialize(&argc,&argv,(char*)0,help);if (ierr) return ierr;
+
+  ierr = PetscOptionsGetInt(NULL,NULL,"-n",&n,NULL);CHKERRQ(ierr);
+  ierr = PetscOptionsGetInt(NULL,NULL,"-m",&m,&flag);CHKERRQ(ierr);
+  if (!flag) m=n;
+  N = n*m;
+  ierr = PetscPrintf(PETSC_COMM_WORLD,"\nModified 2-D Laplacian Eigenproblem, N=%D (%Dx%D grid)\n\n",N,n,m);CHKERRQ(ierr);
+
+  /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+       Compute the operator matrix A and a sparser approximation A0
+     - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
+
+  ierr = MatCreate(PETSC_COMM_WORLD,&A);CHKERRQ(ierr);
+  ierr = MatSetSizes(A,PETSC_DECIDE,PETSC_DECIDE,N,N);CHKERRQ(ierr);
+  ierr = MatSetFromOptions(A);CHKERRQ(ierr);
+  ierr = MatSetUp(A);CHKERRQ(ierr);
+  ierr = MatGetOwnershipRange(A,&Istart,&Iend);CHKERRQ(ierr);
+  for (II=Istart;II<Iend;II++) {
+    i = II/n; j = II-i*n;
+    if (i>0) { ierr = MatSetValue(A,II,II-n,-0.2,INSERT_VALUES);CHKERRQ(ierr); }
+    if (i<m-1) { ierr = MatSetValue(A,II,II+n,-0.2,INSERT_VALUES);CHKERRQ(ierr); }
+    if (j>0) { ierr = MatSetValue(A,II,II-1,-3.0,INSERT_VALUES);CHKERRQ(ierr); }
+    if (j<n-1) { ierr = MatSetValue(A,II,II+1,-3.0,INSERT_VALUES);CHKERRQ(ierr); }
+    ierr = MatSetValue(A,II,II,7.0,INSERT_VALUES);CHKERRQ(ierr);
+  }
+  ierr = MatAssemblyBegin(A,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyEnd(A,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+
+  ierr = MatCreate(PETSC_COMM_WORLD,&A0);CHKERRQ(ierr);
+  ierr = MatSetSizes(A0,PETSC_DECIDE,PETSC_DECIDE,N,N);CHKERRQ(ierr);
+  ierr = MatSetFromOptions(A0);CHKERRQ(ierr);
+  ierr = MatSetUp(A0);CHKERRQ(ierr);
+  ierr = MatGetOwnershipRange(A0,&Istart,&Iend);CHKERRQ(ierr);
+  for (II=Istart;II<Iend;II++) {
+    i = II/n; j = II-i*n;
+    if (j>0) { ierr = MatSetValue(A0,II,II-1,-3.0,INSERT_VALUES);CHKERRQ(ierr); }
+    if (j<n-1) { ierr = MatSetValue(A0,II,II+1,-3.0,INSERT_VALUES);CHKERRQ(ierr); }
+    ierr = MatSetValue(A0,II,II,7.0,INSERT_VALUES);CHKERRQ(ierr);
+  }
+  ierr = MatAssemblyBegin(A0,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyEnd(A0,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+
+  /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+                Create the eigensolver and set various options
+     - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
+
+  ierr = EPSCreate(PETSC_COMM_WORLD,&eps);CHKERRQ(ierr);
+  ierr = EPSSetOperators(eps,A,NULL);CHKERRQ(ierr);
+  ierr = EPSSetProblemType(eps,EPS_HEP);CHKERRQ(ierr);
+  ierr = EPSGetST(eps,&st);CHKERRQ(ierr);
+  ierr = STSetType(st,STSINVERT);CHKERRQ(ierr);
+  ierr = STSetPreconditionerMat(st,A0);CHKERRQ(ierr);
+  ierr = EPSSetTarget(eps,0.0);CHKERRQ(ierr);
+  ierr = EPSSetWhichEigenpairs(eps,EPS_TARGET_MAGNITUDE);CHKERRQ(ierr);
+  ierr = EPSSetFromOptions(eps);CHKERRQ(ierr);
+
+  /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+                 Solve the eigensystem and display solution
+     - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
+
+  ierr = EPSSolve(eps);CHKERRQ(ierr);
+
+  /* show detailed info unless -terse option is given by user */
+  ierr = PetscOptionsHasName(NULL,NULL,"-terse",&terse);CHKERRQ(ierr);
+  if (terse) {
+    ierr = EPSErrorView(eps,EPS_ERROR_RELATIVE,NULL);CHKERRQ(ierr);
+  } else {
+    ierr = PetscViewerPushFormat(PETSC_VIEWER_STDOUT_WORLD,PETSC_VIEWER_ASCII_INFO_DETAIL);CHKERRQ(ierr);
+    ierr = EPSConvergedReasonView(eps,PETSC_VIEWER_STDOUT_WORLD);CHKERRQ(ierr);
+    ierr = EPSErrorView(eps,EPS_ERROR_RELATIVE,PETSC_VIEWER_STDOUT_WORLD);CHKERRQ(ierr);
+    ierr = PetscViewerPopFormat(PETSC_VIEWER_STDOUT_WORLD);CHKERRQ(ierr);
+  }
+  ierr = EPSDestroy(&eps);CHKERRQ(ierr);
+  ierr = MatDestroy(&A);CHKERRQ(ierr);
+  ierr = MatDestroy(&A0);CHKERRQ(ierr);
+  ierr = SlepcFinalize();
+  return ierr;
+}
+
+/*TEST
+
+   testset:
+      args: -eps_nev 4 -terse
+      output_file: output/ex46_1.out
+      requires: !single
+      test:
+         suffix: 1
+      test:
+         suffix: 2
+         args: -st_ksp_type bcgs -st_pc_type bjacobi
+
+TEST*/