+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Import SLEPc4py sources into SLEPc source tree.
 - New tolerance to check a definite inner product, see `BVSetDefiniteTolerance()`.
+- New function `STSetPreconditionerMat()` to pass the matrix for building the
+  preconditioner.
 
 ### Changed
 
@@ -17,6 +19,10 @@
   so setting MUMPS options in the command-line requires the appropriate prefix, e.g.,
   `-st_mat_mumps_icntl_13 1`.
 - `DSTruncate()` now has an additional argument `trim`.
+
+### Deprecated
+
+- `STPrecondSetMatForPC()` is superseded by `STSetPreconditionerMat()`.
 
 ## [3.14] - 2020-09-30